+++ pyi.py
@@ -210,11 +210,8 @@ class PyiVisitor(ast.NodeVisitor):
     def visit_AnnAssign(self, node: ast.AnnAssign) -> None:
         if isinstance(node.annotation, ast.Name) and node.annotation.id == "TypeAlias":
             return
-        if isinstance(node.target, ast.Name):
-            if node.value and not isinstance(node.value, ast.Ellipsis):
-                self.error(node.value, Y015)
-            elif node.value and not self.in_class:
-                self.error(node.value, Y092)
+        if node.value and not isinstance(node.value, ast.Ellipsis):
+            self.error(node.value, Y015)
 
     def _check_union_members(self, members: Sequence[ast.expr]) -> None:
         members_by_dump: dict[str, list[ast.expr]] = {}
@@ -560,7 +557,6 @@ Y015 = 'Y015 Attribute must not have a default value other than "..."'
 Y016 = "Y016 Duplicate union member"
 Y017 = "Y017 Only simple assignments allowed"
 Y018 = 'Y018 {typevarlike_cls} "{typevar_name}" is not used'
-Y092 = "Y092 Top-level attribute must not have a default value"
 Y093 = "Y093 Use typing_extensions.TypeAlias for type aliases"
 
-DISABLED_BY_DEFAULT = [Y092, Y093]
+DISABLED_BY_DEFAULT = [Y093]